fixed invalid QueueItem deletion on exit made workaround to avoid crash in FileChunksInfo::addChunkPos small code changes in UploadQueueItem

+++ windows/TransferView.h
@@ -225,7 +225,7 @@ private:
 
 		static int compareItems(const ItemInfo* a, const ItemInfo* b, uint8_t col);
 
-		int imageIndex() const { return !download ? IMAGE_UPLOAD : (!main ? IMAGE_DOWNLOAD : IMAGE_SEGMENT); }
+		uint8_t imageIndex() const { return static_cast<uint8_t>(!download ? IMAGE_UPLOAD : (!main ? IMAGE_DOWNLOAD : IMAGE_SEGMENT)); }
 
 		ItemInfo* createMainItem() {
 	  		ItemInfo* h = new ItemInfo(user, true);
